Use ProcessingPaymentDialog for lnurl payments as well
Remove ProcessingLNURLPaymentDialog
